Make a copy when authenticating
Rather than for a single request replacing the auth with "",
which may or may not be valid for another request happening at the same time,
replace it to explicitly set the auth to "" only for unauthenticated login request.

This change means that the login code is no longer mutating the local element
in order to log in, thus causing concurrent attempts to not fail with obscure errors.